Move Dealer content representation and route into package
