+++ Packages/EurofurenceApplication/Sources/EurofurenceApplication/Content Router/IntentContentRepresentation.swift
@@ -1,4 +1,5 @@
 import ComponentBase
+import DealerComponent
 import Foundation
 
 public struct IntentContentRepresentation: ContentRepresentation {
