+++ Packages/EurofurenceApplication/Sources/EurofurenceApplication/Content Router/URL Content/URLContentRepresentation.swift
@@ -1,4 +1,5 @@
 import ComponentBase
+import DealerComponent
 import EurofurenceModel
 import Foundation
