+++ Packages/EurofurenceApplication/Tests/EurofurenceApplicationTests/Application/Component Test Doubles/StubDealerDetailComponentFactory.swift
@@ -1,16 +1,1 @@
-import DealerComponent
-import EurofurenceApplication
-import EurofurenceModel
-import UIKit
-import XCTEurofurenceModel
 
-class StubDealerDetailComponentFactory: DealerDetailComponentFactory {
-
-    let stubInterface = UIViewController()
-    private(set) var capturedModel: DealerIdentifier?
-    func makeDealerDetailComponent(for dealer: DealerIdentifier) -> UIViewController {
-        capturedModel = dealer
-        return stubInterface
-    }
-
-}
