+++ Packages/EurofurenceApplication/Tests/EurofurenceApplicationTests/Content Router/IntentContentRepresentationTests.swift
@@ -1,3 +1,4 @@
+import DealerComponent
 import EurofurenceApplication
 import XCTComponentBase
 import XCTest
