+++ Packages/EurofurenceApplication/Tests/EurofurenceApplicationTests/Content Router/Route Tests/Dealers/ShowDealerFromDealersTests.swift
@@ -1,3 +1,4 @@
+import DealerComponent
 import EurofurenceApplication
 import EurofurenceModel
 import XCTComponentBase
