+++ Packages/EurofurenceApplication/Tests/EurofurenceApplicationTests/Content Router/URLContentRepresentationTests.swift
@@ -1,4 +1,5 @@
 import ComponentBase
+import DealerComponent
 import EurofurenceApplication
 import EurofurenceModel
 import XCTComponentBase
