+++ Packages/EurofurenceComponents/Sources/DealerComponent/Routing/DealerContentRepresentation.swift
@@ -1,0 +1,13 @@
+import ComponentBase
+import EurofurenceModel
+import Foundation
+
+public struct DealerContentRepresentation: ContentRepresentation {
+    
+    public var identifier: DealerIdentifier
+    
+    public init(identifier: DealerIdentifier) {
+        self.identifier = identifier
+    }
+    
+}
