+++ Packages/EurofurenceComponents/Sources/DealerComponent/Routing/DealerContentRoute.swift
@@ -1,5 +1,4 @@
 import ComponentBase
-import DealerComponent
 
 public struct DealerContentRoute {
